feat: Update PersistentTabView widget with required arguments
This commit updates the `PersistentTabView` widget in the `persistent_tab_view.widget.dart` file. It adds the required `navBarHeight` argument of type `double` and the `onItemTapped` argument of type `void Function(int index)`. These arguments were missing in the previous implementation. This update ensures that the widget can be used correctly with the required arguments.

Note: Ignoring recent user commits and repository commits as they are unrelated to the code changes.