feat: Add onItemTapped argument to PersistentTabView widget
This commit adds the `onItemTapped` argument of type `void Function(int index)` to the `PersistentTabView` widget in the `persistent_tab_view.widget.dart` file. The `onItemTapped` function will be called when a tab item is tapped, allowing for custom logic to be executed. This enhancement ensures better interactivity and flexibility in using the `PersistentTabView` widget.